+++ src/components/news/component.tsx
@@ -1,52 +1,61 @@
-import React, {useState} from 'react'
-import {render} from "react-dom"
+import React, { useEffect, useState } from 'react';
+import { render } from 'react-dom';
+import axios from 'axios';
 import {
-  NewsApiRequest,
   NewsPageProcessors,
-  NewsArticleContainer,
   MapAPIDataToArticleObject,
   ApiResponseValidator,
   Article,
-} from "../../news/newsFunctionality";
+} from '../../newsData/newsFunctionality';
+import IndividualArticles from './individualArticles';
+import { API_KEY } from './secretKey';
+
+class NewsComponentProcessor {
+  process(undformattedArticles: object[]) {
+    const linkNames: string[] = [];
+
+    const apiResponseValidator = new ApiResponseValidator();
+    let mapAPIDataToArticleObject = new MapAPIDataToArticleObject(
+      apiResponseValidator
+    );
+
+    const article = mapAPIDataToArticleObject.map(undformattedArticles[0]);
+
+    const formattedArticles = [<IndividualArticles article={article} />];
+
+    return formattedArticles;
+  }
+}
 
 export default function NewsComponent() {
-  let newsApiRequest = new NewsApiRequest();
+  const newsProcessor = new NewsComponentProcessor();
 
-  const linkNames: string[] = [];
+  const [apiData, setApiData] = useState([{}]);
 
-  //const apiDataArray = newsApiRequest.apiRequest();
-
-  //const mapAPiDataToArticleObject = new MapAPIDataToArticleObject();
-
-  // for (let apiDataElement of apiDataArray) {
-  //   let article = mapAPiDataToArticleObject.map(apiDataElement);
-  //   linkNames.push(article.name);
-  // }
-
-  const articleObject = {
-    articlesDescription: [
-      { type: "heading", content: "Some heading" },
-      { type: "paragraph", content: "Hello World" },
-    ],
-      articlesName: "name",
-      authors: [{ name: "Alex" }, {name: "Bob"}],
-      dateModified: { date: "10:08", timezone_type: 12, timezone: "UTC" },
-      publishedAt: { date: "10:08", timezone_type: 3, timezone: "UTC" },
+  useEffect(() => {
+    const options = {
+      method: 'GET',
+      url: 'https://reuters-business-and-financial-news.p.rapidapi.com/article-date/2021-04-01',
+      headers: {
+        'X-RapidAPI-Key': API_KEY,
+        'X-RapidAPI-Host': 'reuters-business-and-financial-news.p.rapidapi.com',
+      },
+      timeout: 1000000,
     };
 
-  const apiResponseValidator = new ApiResponseValidator();
-  let mapAPIDataToArticleObject = new MapAPIDataToArticleObject(apiResponseValidator);
+    axios
+      .request(options)
+      .then((response) => {
+        setApiData(response.data);
+      })
+      .catch((e) => console.error(e));
+  }, []);
 
-  const article = mapAPIDataToArticleObject.map(articleObject);
+  const articles = newsProcessor.process(apiData);
 
-  const newsArticleContainerObject = new NewsArticleContainer();
-  let newsArticleContainer = newsArticleContainerObject.create(article);
-
-  const articles = [
-    newsArticleContainer
-  ]
-
-  return (<div>{articles}</div>);
+  return <div>{articles}</div>;
 }
 
 // https://codingmanatee.wordpress.com/2023/07/12/type-number-is-not-assignable-to-type-reactnode-ts2322/ Used to figure out that I need to create a string for JSX
+// https://stackoverflow.com/questions/65277539/property-props-does-not-exist-on-type-intrinsicattributes-string
+// https://www.youtube.com/watch?v=ZEKBDXGnD4s&t=650s Used for useEffect
